♻️ engineering_conversation → agents 이동 + discord compat shim
변경 이유:
discord/engineering_conversation 도 discord transport 가 없는
engineering-agent task-shaping/intent/research-bootstrap 로직이라
agents↔discord 순환의 다른 축을 만든다. agents/ 로 옮긴다.

주요 변경 사항:
- src/yule_orchestrator/discord/engineering_conversation → agents/ 로 relocate
  (6개 책임 모듈 + facade).
- 내부 cross-package import 절대경로화 (from ...agents → from yule_orchestrator.agents),
  discord sibling 참조 (from ..engineering.help_surface) →
  from yule_orchestrator.discord.engineering.help_surface 절대경로화.
- 기존 discord 경로에 compat shim: facade __init__ public(+__all__ 외
  _suggest_task_type) 재노출, 6개 submodule 각각 sys.modules alias 로
  identity 보존 (patch 대상 유지).

비고:
shim 으로 discord→agents forward 방향만 남긴다.
